% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{bm_kl}
\alias{bm_kl}
\title{Generate brownian motion using the KL decomposition}
\usage{
bm_kl(k, n)
}
\arguments{
\item{k}{Numeric, number of basis functions.}

\item{n}{Numeric, number of sampling points.}
}
\value{
List, containing the sampling points, observed points, and normalised scores.
}
\description{
Generate brownian motion using the KL decomposition
}
