% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{bm_kl_rd}
\alias{bm_kl_rd}
\title{Generate brownian motion using the KL decomposition on a random grid of points}
\usage{
bm_kl_rd(k, x)
}
\arguments{
\item{k}{Numeric, number of basis functions.}

\item{x}{Vector containing the sampling points.}
}
\value{
List, containing the sampling points, observed points, and normalised scores.
}
\description{
Generate brownian motion using the KL decomposition on a random grid of points
}
