% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{boat_cdf}
\alias{boat_cdf}
\title{Generate points from a convex density function}
\usage{
boat_cdf(a, n)
}
\arguments{
\item{a}{Numeric, multiplicative parameter.}

\item{n}{Numeric, number of points to generate}
}
\value{
Vector of sample points.
}
\description{
Points are generated using the inverse transform sampling method.
}
