% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{bs_kl}
\alias{bs_kl}
\title{Simulation brownian sheet using the Kosambi-Karhunen-Loève decomposition}
\usage{
bs_kl(xout, k, nu1 = 1, nu2 = 1)
}
\arguments{
\item{xout}{Data frame, consisting of:
\itemize{
\item \strong{$t1} Vector containing the coordinates in the first dimension.
\item \strong{$t2} Vector containing the coordinates in the second dimension.
}}

\item{k}{Numeric, the number of basis functions.}
}
\value{
List, containing:
\itemize{
\item \strong{$x_obs} Data frame containing the observed points at the \verb{(t1, t2)}
coordinates.
\item \strong{$xi} Matrix containing the scores.
}
}
\description{
Simulation brownian sheet using the Kosambi-Karhunen-Loève decomposition
}
