% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{cdf_series}
\alias{cdf_series}
\title{Estimates the cumulative distribution function using series expansions}
\usage{
cdf_series(xi, x, J, thresh = TRUE, cj0 = 3, cj1 = 0.8, cth = 4)
}
\arguments{
\item{xi}{Vector of observed points.}

\item{x}{Vector of evaluation points.}

\item{J}{Numeric, number of basis functions. Ignored if thresholding
is used.}

\item{thresh}{Boolean, indicating whether thresholding should be performed.}

\item{cj0}{Numeric.}

\item{cj1}{Numeric.}

\item{cth}{Numeric.}
}
\value{
Vector, the estimated cdf.
}
\description{
Estimates the cumulative distribution function using series expansions
}
