% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{class_score}
\alias{class_score}
\title{Computes metric for binary classifiers}
\usage{
class_score(y_hat, y_true, type = "f1")
}
\arguments{
\item{y_hat}{Vector, the predicted classes.}

\item{y_true}{Vector, the true classes.}

\item{type}{String, either \code{f1} or \code{accuracy}.}
}
\value{
Numeric.
}
\description{
Computes metric for binary classifiers
}
