% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{confint_lim}
\alias{confint_lim}
\title{Computes the confidence intervals for scores based on the asymptotic limit
of leave-one-out control neighbours}
\usage{
confint_lim(scores_list, pdf_list, conf_level)
}
\arguments{
\item{scores_list}{List, containing the elements:
\itemize{
\item \strong{$weights} Vector, containing the weights associated to each score.
\item \strong{$xi_hat} Vector, containing the estimated scores.
\item \strong{$varphi} Matrix, containing the evaluated functional, with rows carrying
the sampling points and columns carrying the components.
}}

\item{pdf_list}{List, containing the elements:
\itemize{
\item \strong{t} Vector containing the sampling points.
\item \strong{x} Vector containing the observed points.
}}

\item{conf_level}{Numeric, indicating the confidence level in which the
intervals should be constructed.}
}
\value{
List, containing the elements:
\itemize{
\item \strong{$scores} Vector, containing the estimated scores.
\item \strong{$ci_l} Vector, containing the lower bound of the confidence interval.
\item \strong{$ci_u} Vector, containing the upper bound of the confidence interval.
\item \strong{$sm} Vector, containing the standard deviation.
}
}
\description{
Computes the confidence intervals for scores based on the asymptotic limit
of leave-one-out control neighbours
}
