% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint_pace}
\alias{confint_pace}
\title{Compute confidence intervals based on the PACE method}
\usage{
confint_pace(scores_vec, lambda_vec, phi_list, Sigma_mat, conf_level)
}
\arguments{
\item{scores_vec}{Vector containing the scores.}

\item{lambda_vec}{Vector containing the eigenvalues.}

\item{phi_list}{List of the eigenfunctions, containing the elements:
\itemize{
\item \strong{$t} Vector of sampling points.
\item \strong{$X} Matrix of observed points, with sampling points on rows and the index.
}}

\item{Sigma_mat}{Matrix containing the covariance plus noise.}

\item{conf_level}{Numeric, indicating the confidence level in which the
intervals should be constructed.}
}
\value{
List, containing the elements:
\itemize{
\item \strong{$scores} Vector, containing the estimated scores.
\item \strong{$ci_l} Vector, containing the lower bound of the confidence interval.
\item \strong{$ci_u} Vector, containing the upper bound of the confidence interval.
\item \strong{$sm} Vector, containing the standard deviation.
}
}
\description{
Compute confidence intervals based on the PACE method
}
