% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{cos_basis}
\alias{cos_basis}
\title{Constructs an orthonormal cosine basis}
\usage{
cos_basis(x, J)
}
\arguments{
\item{x}{Vector of evaluation points.}

\item{J}{Numeric, the number of basis functions to take. Should range
between 0 and an integer value, where 0 indicates only the constant function
to be taken.}
}
\value{
Matrix of dimension \code{x} by \code{J + 1}.
}
\description{
Constructs an orthonormal cosine basis
}
