% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{cum_vol}
\alias{cum_vol}
\title{Calculates the leave-one-out cumulative volume for general distributions}
\usage{
cum_vol(x, cdf = identity)
}
\arguments{
\item{x}{Vector of points.}

\item{cdf}{Function, representation the empirical distribution function.}
}
\value{
List, containing the associated leave-one-out cumulative volume
for each sampling point.
}
\description{
Given the distribution function and a vector of points, the leave-one-out
cumulative volume in one dimension is calculated based on order statistics.
}
