% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{cum_vol_unif}
\alias{cum_vol_unif}
\title{Calculates the leave-one-out cumulative volume for the uniform distribution}
\usage{
cum_vol_unif(x)
}
\arguments{
\item{x}{Vector of points.}
}
\value{
List, containing the points and its associated cumulative
volume.
}
\description{
Calculates the leave-one-out cumulative volume for the uniform distribution
}
