% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{degree}
\alias{degree}
\title{Calculates the degree of a vector}
\usage{
degree(x)
}
\arguments{
\item{x}{Vector of points.}
}
\value{
List, containing the associated degree for each sampling
point.
}
\description{
For each point in a vector, the number of points in which it is
a nearest neighbour to is calculated.
}
