% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{den_series}
\alias{den_series}
\title{Estimate the density using a series expansion}
\usage{
den_series(xi, x, J, thresh = TRUE, cj0 = 3, cj1 = 0.8, cth = 4)
}
\arguments{
\item{xi}{Vector, the observed design points.}

\item{x}{Vector, the evaluation points to estimate the density.}

\item{J}{Numeric, the number of basis functions.}
}
\value{
Vector, the estimated density on the vector of evaluation points.
}
\description{
Estimate the density using a series expansion
}
