% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{int_cos_basis}
\alias{int_cos_basis}
\title{Construct the integrated cosine basis}
\usage{
int_cos_basis(x, J)
}
\arguments{
\item{x}{Vector of evaluation points.}

\item{J}{Numeric, number of basis functions.}
}
\value{
Matrix, containing the basis functions on the evaluation
points.
}
\description{
Used for the estimation for the cumulative distribution function.
}
