% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_nn.R
\name{mc_ci}
\alias{mc_ci}
\title{Computes confidence intervals with control neighbours integration method in
the presence of measurement errors}
\usage{
mc_ci(varphi_noise, conf_level = 0.95, lim = FALSE)
}
\arguments{
\item{varphi_noise}{List, containing the following elements:
\itemize{
\item \strong{$weights} Vector of control neighbour weights.
\item \strong{$varphi_int} Numeric, the estimated integral value.
\item \strong{$varphi_noise} Vector, the noise functional in which the
confidence intervals should be based on.
}}

\item{conf_level}{Numeric, the desired confidence level.}

\item{lim}{Boolean, indicating whether to use the asymptotic confidence intervals.
If FALSE, the non-asymptotic version based on conditional variance is used.}
}
\value{
List, containing the following elements:
\itemize{
\item \strong{$varphi_int} Numeric, the estimated integral value.
\item \strong{$ci_l} Numeric, the lower bound of the confidence interval.
\item \strong{$ci_u} Numeric, the upper bound of the confidence interval.
}
}
\description{
Computes confidence intervals with control neighbours integration method in
the presence of measurement errors
}
