% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_nn.R
\name{mc_int}
\alias{mc_int}
\title{Estimates the integral of a function using control neighbours method}
\usage{
mc_int(x, varphi, cdf)
}
\arguments{
\item{x}{Vector of sampling points.}

\item{varphi}{Vector containing the observed points of the integrand.}

\item{cdf}{List or a function representing the cumulative distribution function.
If list, it should contain the elements:
\itemize{
\item \strong{$t} Vector of sampling points.
\item \strong{$x} Vector of observed points.
}}
}
\value{
Numeric, the estimated integral.
}
\description{
Estimates the integral of a function using control neighbours method
}
