% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_nn.R
\name{mc_pi}
\alias{mc_pi}
\title{Computes prediction intervals based on control neighbours based on bootstrap}
\usage{
mc_pi(
  varphi,
  varphi_int,
  eps = 0.05,
  s,
  b_out = 200,
  d = 1,
  cdf = identity,
  pow
)
}
\arguments{
\item{varphi}{Vector, the integrand on a vector of evaluation points.}

\item{varphi_int}{Numeric, the integral estimated based on control neighbours.}

\item{eps}{Numeric, the error critical value.}

\item{s}{Numeric, the Hölder exponent of the integrand.}

\item{b_out}{Numeric, number of bootstrap replications to perform.}

\item{d}{Numeric, the dimension of the design points. Only one dimension is
implemented at the moment.}

\item{cdf}{Function, indicating the cumulative distribution function of the
design points.}
}
\value{
List, containing the following elements:
\itemize{
\item \strong{$varphi_int} Numeric, the estimated integral value.
\item \strong{$pi_l} Numeric, the lower bound of the prediction interval.
\item \strong{$pi_u} Numeric, the upper bound of the prediction interval.
\item \strong{$width} Numeric, the width of the prediction interval.
}
}
\description{
Computes prediction intervals based on control neighbours based on bootstrap
}
