% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pace_scores}
\alias{pace_scores}
\title{Compute the scores using conditional expectation of PACE}
\usage{
pace_scores(y_list, lambda_vec, phi_list, Sigma_mat, mu_list)
}
\arguments{
\item{y_list}{List of the observations, containing the elements:
\itemize{
\item \strong{$t} Vector of sampling points.
\item \strong{$y} Vector of observed points.
}}

\item{lambda_vec}{Vector containing the eigenvalues.}

\item{phi_list}{List of the eigenfunctions, containing the elements:
\itemize{
\item \strong{$t} Vector of sampling points.
\item \strong{$X} Matrix of observed points, with sampling points on rows and the index
}}

\item{Sigma_mat}{Matrix containing the covariance plus noise.}

\item{mu_list}{List of the mean function, containing the elements:
\itemize{
\item \strong{$t} Vector of sampling points.
\item \strong{$x} Vector of observed points.
}}
}
\value{
Vector containing the estimated scores.
}
\description{
Compute the scores using conditional expectation of PACE
}
