% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_nn.R
\name{pi_subsam2d}
\alias{pi_subsam2d}
\title{Estimates prediction intervals using subsampling for 2 dimensions.}
\usage{
pi_subsam2d(X, X_int, eps = 0.05, s, b_out = 200, int_fun)
}
\arguments{
\item{X}{Data frame, containing the following columns:
\itemize{
\item \strong{$t1} Vector, containing the coordinates in \code{t1}.
\item \strong{$t2} Vector, containing the coordinates in \code{t2}.
\item \strong{$x} Vector of observed points, where each point is associated with each
row of coordinates in \code{t}.
}}

\item{X_int}{Numeric, the control neighbour integral estimate.}

\item{eps}{Numeric, the error critical value.}

\item{s}{Numeric, the Hölder exponent of the integrand.}

\item{b_out}{Numeric, number of bootstrap replications to perform.}

\item{int_fun}{Function, indicating the method of approximating integrals.
Only \code{mean} and \code{riemann_2d} function is accepted.}
}
\value{
List, containing the following elements:
\itemize{
\item \strong{$X_int} Numeric, the estimated integral value.
\item \strong{$pi_l} Numeric, the lower bound of the prediction interval.
\item \strong{$pi_u} Numeric, the upper bound of the prediction interval.
\item \strong{$width} Numeric, the width of the prediction interval.
}
}
\description{
Mainly used for comparison purposes. Only works with either the sample mean
or "riemann sums" using voronoi cells.
}
