% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_nn.R
\name{riemann_2d}
\alias{riemann_2d}
\title{Compute Riemann sum estimates based on Voronoi partition}
\usage{
riemann_2d(X, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
}
\arguments{
\item{X}{Data frame, containing the following columns:
-\strong{$t1} Vector, containing the coordinates in \code{t1}.
-\strong{$t2} Vector, containing the coordinates in \code{t2}.
-\strong{$x} Vector of observed points, where each point is associated with each
row of coordinates in \code{t}.}

\item{xmin}{Numeric, the minimum coordinate of the \code{x} axis.}

\item{xmax}{Numeric, the maximum coordinate of the \code{x} axis.}

\item{ymin}{Numeric, the minimum coordinate of the \code{y} axis.}

\item{ymax}{Numeric, the maximum coordinate of the \code{y} axis.}
}
\value{
Numeric, the estimated integral.
}
\description{
Compute Riemann sum estimates based on Voronoi partition
}
