% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{splines_mean}
\alias{splines_mean}
\title{Estimate the mean function of functional data by pooling and smoothing splines}
\usage{
splines_mean(X_list, x)
}
\arguments{
\item{X_list}{List, containing the following elements:
-\strong{$t} Vector of observed sampling points.
-\strong{$x} Vector of observed points.}

\item{x}{Vector of evaluation points.}
}
\value{
List, containing the mean function and its evaluation points
}
\description{
Estimate the mean function of functional data by pooling and smoothing splines
}
