% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_loo}
\alias{weights_loo}
\title{Compute weights for leave-one-out control neighbours}
\usage{
weights_loo(x, cdf = identity)
}
\arguments{
\item{x}{Vector of sampling points.}

\item{cdf}{Function, representing the cumulative distribution function.}
}
\value{
List, containing the associated weights for each sampling point.
}
\description{
Compute weights for leave-one-out control neighbours
}
