% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_loo_unif}
\alias{weights_loo_unif}
\title{Calculates the leave-one-out-weights for the uniform distribution}
\usage{
weights_loo_unif(x)
}
\arguments{
\item{x}{Vector of points.}
}
\value{
List, containing the points and asssociated weights.
}
\description{
Calculates the leave-one-out-weights for the uniform distribution
}
