% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{confint_mc}
\alias{confint_mc}
\title{Computes the confidence intervals for scores based on the conditional variance
of leave-one-out control neighbours}
\usage{
confint_mc(scores_list, conf_level)
}
\arguments{
\item{scores_list}{List, containing the elements:
-\strong{$weights} Vector, containing the weights associated to each score.
-\strong{$xi_hat} Vector, containing the estimated scores.
-\strong{$varphi} Matrix, containing the evaluated functional, with rows carrying
the sampling points and columns carrying the components.}

\item{conf_level}{Numeric, indicating the confidence level in which the
intervals should be constructed.}
}
\value{
List, containing the elements:
-\strong{$scores} Vector, containing the estimated scores.
-\strong{$ci_l} Vector, containing the lower bound of the confidence interval.
-\strong{$ci_u} Vector, containing the upper bound of the confidence interval.
}
\description{
Computes the confidence intervals for scores based on the conditional variance
of leave-one-out control neighbours
}
